Split input
