+++ src/ops/delete_bucket.rs
@@ -1,7 +1,15 @@
 //! [`DeleteBucket`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_DeleteBucket.html)
 
 use super::*;
-use crate::dto::{DeleteBucketError, DeleteBucketOutput};
+use crate::dto::{DeleteBucketError, DeleteBucketOutput, DeleteBucketRequest};
+
+/// extract operation request
+pub fn extract(bucket: &str) -> Result<DeleteBucketRequest, BoxStdError> {
+    let input: DeleteBucketRequest = DeleteBucketRequest {
+        bucket: bucket.into(),
+    };
+    Ok(input)
+}
 
 impl S3Output for DeleteBucketOutput {
     fn try_into_response(self) -> S3Result<Response> {
