+++ src/ops/get_bucket_location.rs
@@ -1,7 +1,15 @@
 //! [`GetBucketLocation`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_GetBucketLocation.html)
 
 use super::*;
-use crate::dto::{GetBucketLocationError, GetBucketLocationOutput};
+use crate::dto::{GetBucketLocationError, GetBucketLocationOutput, GetBucketLocationRequest};
+
+/// extract operation request
+pub fn extract(bucket: &str) -> Result<GetBucketLocationRequest, BoxStdError> {
+    let input = GetBucketLocationRequest {
+        bucket: bucket.into(),
+    };
+    Ok(input)
+}
 
 impl S3Output for GetBucketLocationOutput {
     fn try_into_response(self) -> S3Result<Response> {
