+++ src/ops/head_bucket.rs
@@ -1,7 +1,15 @@
 //! [`HeadBucket`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_HeadBucket.html)
 
 use super::*;
-use crate::dto::{HeadBucketError, HeadBucketOutput};
+use crate::dto::{HeadBucketError, HeadBucketOutput, HeadBucketRequest};
+
+/// extract operation request
+pub fn extract(bucket: &str) -> Result<HeadBucketRequest, BoxStdError> {
+    let input = HeadBucketRequest {
+        bucket: bucket.into(),
+    };
+    Ok(input)
+}
 
 impl S3Output for HeadBucketOutput {
     fn try_into_response(self) -> S3Result<Response> {
