+++ src/ops/put_object.rs
@@ -1,7 +1,37 @@
 //! [`PutObject`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_PutObject.html)
 
 use super::*;
-use crate::dto::{PutObjectError, PutObjectOutput};
+use crate::dto::{ByteStream, PutObjectError, PutObjectOutput, PutObjectRequest};
+use futures::stream::StreamExt;
+use std::io;
+
+/// extract operation request
+pub fn extract(
+    _req: &Request,
+    body: Body,
+    bucket: &str,
+    key: &str,
+) -> Result<PutObjectRequest, BoxStdError> {
+    let body = body
+        .map(|try_chunk| {
+            try_chunk.map(|c| c).map_err(|e| {
+                io::Error::new(
+                    io::ErrorKind::Other,
+                    format!("Error obtaining chunk: {}", e),
+                )
+            })
+        })
+        .apply(ByteStream::new);
+
+    let input: PutObjectRequest = PutObjectRequest {
+        bucket: bucket.into(),
+        key: key.into(),
+        body: Some(body),
+        ..PutObjectRequest::default() // TODO: handle other fields
+    };
+
+    Ok(input)
+}
 
 impl S3Output for PutObjectOutput {
     fn try_into_response(self) -> S3Result<Response> {
